Update bounding box to include all of Raleigh
The current bounding box for Raleigh seems to include Durham and Chapel Hill, but only some of Raleigh. As such, it appears that what the Raleigh extract is meant to encompass the broader Triangle regions. This updated bounding box for Raleigh ensures that all of Raleigh is included in the extract while also returning all of Durham and Chapel Hill.